refactor: divides AudioPlayer into separate components
