+++ src/App.vue
@@ -3,19 +3,10 @@
     <div class="container">
       <h4>multiple audio example</h4>
       <div v-for="(audio, idx) of audios" :key="idx">
-        <!-- <AudioPlayer
-          v-if="idx === 1" ref="audioPlayer"
-          :src="audio.src" :idx="idx" :stream="audio.stream" :volume-bar="audio.volumeBar"
-          :audio-status="audio.status" :data-tracking="audio.dataTracking" :hidden="audio.hidden" :master-volume="0.75"
-          @error="handleError($event, idx)" @audio-status-updated="updateAudioStatus" @amplitude-data="onAmplitudeData"
-        /> -->
-        <AudioPlayer
-          ref="audioPlayer" :src="audio.src" :idx="idx" :stream="audio.stream" :volume-bar="audio.volumeBar"
-          :audio-status="audio.status" :data-tracking="audio.dataTracking" :hidden="audio.hidden" :master-volume="0.75"
-          @error="handleError($event, idx)" @audio-status-updated="updateAudioStatus" @amplitude-data="onAmplitudeData"
-        >
+        <AudioStreamPlayer v-if="audio.stream" ref="audioPlayer" :src="audio.src" :idx="idx" :volume-bar="audio.volumeBar" :audio-status="audio.status" :data-tracking="audio.dataTracking" :hidden="audio.hidden" :master-volume="0.75" @error="handleError($event, idx)" @audio-status-updated="updateAudioStatus" @amplitude-data="onAmplitudeData">
           <div v-if="audio.dataTracking" class="amplitude data-tracking" :style="{ background: audio.dataTracking && audio.status === 'playing' ? `rgb(199 0 57 / ${dbOpacity}%` : 'transparent', transform: `scale(${dbOpacity / 100 * 2})` }" />
-        </AudioPlayer>
+        </AudioStreamPlayer>
+        <AudioFilePlayer v-else ref="audioPlayer" :src="audio.src" :idx="idx" :stream="audio.stream" :audio-status="audio.status" :hidden="audio.hidden" :master-volume="0.75" @error="handleError($event, idx)" @audio-status-updated="updateAudioStatus" />
         <button v-if="audio?.status !== 'error'" :disabled="!audio.status" @click="toggleAudio(idx)">
           {{ audio.status ? audio.status : 'initializing' }}
         </button>
@@ -26,15 +17,9 @@
     </div>
     <div class="container">
       <h4>single audio example</h4>
-      <AudioPlayer
-        :src="audioFile" :audio-status="audioStatus" :play-on-mount="playOnMount"
-        @audio-status-updated="updateAudioStatus"
-      >
-        <ExtendedInfo
-          :audio-data="audioData" :extended-info-open="false" :queue-button="true"
-          :location-button="true"
-        />
-      </AudioPlayer>
+      <AudioFilePlayer :src="audioFile" :audio-status="audioStatus" :play-on-mount="playOnMount" @audio-status-updated="updateAudioStatus">
+        <ExtendedInfo :audio-data="audioData" :extended-info-open="false" :queue-button="true" :location-button="true" />
+      </AudioFilePlayer>
       <button @click="toggleAudio()">
         {{ audioStatus }}
       </button>
@@ -50,20 +35,12 @@
     </div>
     <div class="container">
       <h4>single audio example 2</h4>
-      <AudioPlayer
-        :src="audioFile" :audio-status="audioStatus" :next-button="true" :previous-button="true"
-        :play-on-mount="playOnMount" :volume-button="false" :shuffle-button="true" :spacebar-toggle="true" rounded
-        :extended-info-opened="showExtended" @audio-status-updated="updateAudioStatus" @next="handleNext"
-        @previous="handlePrevious" @shuffle-toggle="handleShuffleToggle"
-      >
+      <AudioFilePlayer :src="audioFile" :audio-status="audioStatus" :next-button="true" :previous-button="true" :play-on-mount="playOnMount" :volume-button="false" :shuffle-button="true" :spacebar-toggle="true" rounded :extended-info-opened="showExtended" @audio-status-updated="updateAudioStatus" @next="handleNext" @previous="handlePrevious" @shuffle-toggle="handleShuffleToggle">
         <template v-if="showExtended" #extended-top>
           <div>test</div>
         </template>
-        <ExtendedInfo
-          :audio-data="audioData" :extended-info-open="false" :location-button="true"
-          @extended-click="showExtended = !showExtended"
-        />
-      </AudioPlayer>
+        <ExtendedInfo :audio-data="audioData" :extended-info-open="false" :location-button="true" @extended-click="showExtended = !showExtended" />
+      </AudioFilePlayer>
       <button @click="toggleAudio()">
         {{ audioStatus }}
       </button>
@@ -82,7 +59,7 @@
 
 <script lang="ts" setup>
 import { ref, useTemplateRef } from 'vue'
-import AudioPlayer from './components'
+import { AudioFilePlayer, AudioStreamPlayer } from './components'
 import ExtendedInfo from './components/ExtendedInfo.vue'
 
 const audioPlayer = useTemplateRef('audioPlayer')
@@ -164,7 +141,7 @@ function handleSeek(idx: number, time: number) {
 }
 
 function updateAudioStatus(status: string, idx: number) {
-  console.log({ status })
+  console.log({ status, idx })
   if (idx !== null) {
     audios.value[idx].status = status
   } else {
@@ -219,9 +196,10 @@ function onAmplitudeData(data: any) {
 .content {
   font-family: AuthenticSans, Arial, sans-serif;
   color: red;
+  padding: 1rem 0 2rem;
 
   .container {
-    padding: 1rem 0;
+    padding: 0;
 
     h4 {
       padding: 1rem 1rem 0;
