+++ src/assets/scss/_fonts.scss
@@ -1,5 +1,3 @@
-@import 'ress';
-
 @font-face {
   font-family: SpaceGrotesk;
   src: url('../assets/fonts/space-grotesk/SpaceGrotesk-Regular.woff2') format('woff2');
