+++ src/assets/scss/global.scss
@@ -1,2 +1,3 @@
 @import 'variables';
 @import 'mixins';
+@import 'fonts';
